Improve performance by reusing python builtin hashing of string dict
As mentionned in a comment on issue #31 [1], the creating of a custom
hashing algorithm is too slow when compared to the native string hashing
algorithm of python. Instead of having a dictionary of ints, we just keep
a dictionary of strings and this has no impact on functionality but improves
the overall performance of loading a dictionary by about 40%

This also adds a 'data_version' key to the pickle dictionary, which starts
at 2 and load_pickle returns False if the data version is incompatible.

[1] https://github.com/mammothb/symspellpy/issues/31#issuecomment-467311165