Adding casing transfer as post-processing to lookup() and lookup_compound()
I have added a post-processing step allowing us to transfer the casing (eg lowercase / uppercase) to the typo-fixed text.
It uses the difflib module's SequenceMatcher to identify the differences between the original phrase and the typo-fixed one and then we carry over the casing based on the type of operation required to transform text1 to text2.